Renamed conntroller closses.py
