+++ src/simulator_core/entities/assets/controller/controller_abstract_clas.py
