+++ src/simulator_core/entities/assets/controller/controller_producer.py
@@ -16,7 +16,8 @@
 
 from simulator_core.entities.assets.asset_defaults import (DEFAULT_TEMPERATURE,
                                                            DEFAULT_TEMPERATURE_DIFFERENCE)
-from simulator_core.entities.assets.controller.controller_classes import AssetControllerAbstract
+from simulator_core.entities.assets.controller.controller_abstract_clas import (
+    AssetControllerAbstract)
 from simulator_core.entities.assets.esdl_asset_object import EsdlAssetObject
